// src/common/enums/permission.enum.ts
export enum Permission {
  MANAGE_USERS = 'MANAGE_USERS',
  REVIEW_PROJECTS = 'REVIEW_PROJECTS',
  APPROVE_PROJECTS = 'APPROVE_PROJECTS',
  MANAGE_PROJECTS = 'MANAGE_PROJECTS',
  MANAGE_TREASURY = 'MANAGE_TREASURY',
  SUBMIT_PROJECT = 'SUBMIT_PROJECT',
  EDIT_OWN_PROJECT = 'EDIT_OWN_PROJECT',
  INVEST = 'INVEST',
  MANAGE_ESCROW = 'MANAGE_ESCROW',
}